+++ tests/ui/apps/WizardPanels/WizardPanels-View.js
@@ -1,7 +1,7 @@
 import BodyText from '../../../../BodyText/BodyText';
 import Button from '../../../../Button/Button';
 import Scroller from '../../../../Scroller/Scroller';
-import {WizardPanel, View} from '../../../../Panels/WizardPanel';
+import WizardPanels from '../../../../WizardPanels';
 import ThemeDecorator from '../../../../ThemeDecorator/ThemeDecorator';
 import React from 'react';
 import spotlight from '@enact/spotlight';
@@ -11,13 +11,13 @@ import spotlight from '@enact/spotlight';
 spotlight.setPointerMode(false);
 
 const app = (props) => <div {...props}>
-	<WizardPanel
-		id="wizardpanel"
+	<WizardPanels
+		id="wizardpanels"
 		nextButtonText="next"
 		prevButtonText="previous"
-		title="WizardPanel"
+		title="WizardPanels"
 	>
-		<View footer="Footer in View 1" subtitle="A subtitle for View 1" title="WizardPanel View 1">
+		<WizardPanels.Panel footer="Footer in View 1" subtitle="A subtitle for View 1" title="WizardPanels View 1">
 			<Scroller>
 				<div id="view1" />
 				<BodyText>Lorem ipsum dolor sit amet, consectetur adipiscing elit, sed do eiusmod tempor incididunt ut labore et dolore magna aliqua. Purus faucibus ornare suspendisse sed nisi. Vestibulum sed arcu non odio euismod lacinia at quis. Elementum eu facilisis sed odio morbi quis commodo. Scelerisque mauris pellentesque pulvinar pellentesque habitant morbi. Neque ornare aenean euismod elementum. Iaculis nunc sed augue lacus viverra vitae congue eu consequat. Vulputate eu scelerisque felis imperdiet proin fermentum leo vel orci. Tincidunt augue i\nterdum velit euismod. Nunc sed augue lacus viverra vitae congue eu consequat. Ultricies integer quis auctor elit sed vulputate. Pellentesque adipiscing commodo elit at imperdiet dui accumsan sit. Elit sed vulputate mi sit amet mauris commodo. Ipsum consequat nisl vel pretium lectus. Sed ullamcorper morbi tincidunt ornare massa eget egestas. Nulla facilisi morbi tempus iaculis urna id volutpat. Facilisis magna etiam tempor orci eu lobortis elementum nibh tellus. Lorem ipsum dolor sit amet, consectetur adipiscing elit, sed do eiusmod tempor incididunt ut labore et dolore magna aliqua. Purus faucibus ornare suspendisse sed nisi. Vestibulum sed arcu non odio euismod lacinia at quis. Elementum eu facilisis sed odio morbi quis commodo. Scelerisque mauris pellentesque pulvinar pellentesque habitant morbi. Neque ornare aenean euismod elementum. Iaculis nunc sed augue lacus viverra vitae congue eu consequat. Vulputate eu scelerisque felis imperdiet proin fermentum leo vel orci. Tincidunt augue interdum velit euismod. Nunc sed augue lacus viverra vitae congue eu consequat. Ultricies integer quis auctor elit sed vulputate. Pellentesque adipiscing commodo elit at imperdiet dui accumsan sit. Elit sed vulputate mi sit amet mauris commodo. Ipsum consequat nisl vel pretium lectus. Sed ullamcorper morbi tincidunt ornare massa eget egestas. Nulla facilisi morbi tempus iaculis urna id volutpat. Facilisis magna etiam tempor orci eu lobortis elementum nibh tellus.Lorem ipsum dolor sit amet, consectetur adipiscing elit, sed do eiusmod tempor incididunt ut labore et dolore magna aliqua. Purus faucibus ornare suspendisse sed nisi. Vestibulum sed arcu non odio euismod lacinia at quis. Elementum eu facilisis sed odio morbi quis commodo. Scelerisque mauris pellentesque pulvinar pellentesque habitant morbi. Neque ornare aenean euismod elementum. Iaculis nunc sed augue lacus viverra vitae congue eu consequat. Vulputate eu scelerisque felis imperdiet proin fermentum leo vel orci. Tincidunt augue interdum velit euismod. Nunc sed augue lacus viverra vitae congue eu consequat. Ultricies integer quis auctor elit sed vulputate. Pellentesque adipiscing commodo elit at imperdiet dui accumsan sit. Elit sed vulputate mi sit amet mauris commodo. Ipsum consequat nisl vel pretium lectus. Sed ullamcorper morbi tincidunt ornare massa eget egestas. Nulla facilisi morbi tempus iaculis urna id volutpat. Facilisis magna etiam tempor orci eu lobortis elementum nibh tellus. Lorem ipsum dolor sit amet, consectetur adipiscing elit, sed do eiusmod tempor incididunt ut labore et dolore magna aliqua. Purus faucibus ornare suspendisse sed nisi. Vestibulum sed arcu non odio euismod lacinia at quis. Elementum eu facilisis sed odio morbi quis commodo. Scelerisque mauris pellentesque pulvinar pellentesque habitant morbi. Neque ornare aenean euismod elementum. Iaculis nunc sed augue lacus viverra vitae congue eu consequat. Vulputate eu scelerisque felis imperdiet proin fermentum leo vel orci. Tincidunt augue interdum velit euismod. Nunc sed augue lacus viverra vitae congue eu consequat. Ultricies integer quis auctor elit sed vulputate. Pellentesque adipiscing commodo elit at imperdiet dui accumsan sit. Elit sed vulputate mi sit amet mauris commodo. Ipsum consequat nisl vel pretium lectus. Sed ullamcorper morbi tincidunt ornare massa eget egestas. Nulla facilisi morbi tempus iaculis urna id volutpat. Facilisis magna etiam tempor orci eu lobortis elementum nibh tellus.</BodyText>
@@ -26,8 +26,8 @@ const app = (props) => <div {...props}>
 				<Button>OK</Button>
 				<Button>Cancel</Button>
 			</buttons>
-		</View>
-		<View subtitle="A subtitle for View 2" title="WizardPanel View 2">
+		</WizardPanels.Panel>
+		<WizardPanels.Panel subtitle="A subtitle for View 2" title="WizardPanels View 2">
 			<div id="view2" />
 			<BodyText>Several buttons!</BodyText>
 			<Button icon="list">Button A</Button>
@@ -41,24 +41,24 @@ const app = (props) => <div {...props}>
 			<footer>
 				<Button inline>Confirm</Button>
 			</footer>
-		</View>
-		<View ooter="Footer in View 3" subtitle="A subtitle for View 3" title="WizardPanel View 3">
+		</WizardPanels.Panel>
+		<WizardPanels.Panel ooter="Footer in View 3" subtitle="A subtitle for View 3" title="WizardPanels View 3">
 			<div id="view3" />
 			<BodyText>A simple view</BodyText>
 			<buttons>
 				<Button>OK</Button>
 				<Button>Cancel</Button>
 			</buttons>
-		</View>
-		<View footer="Footer in View 4" subtitle="A subtitle for View 4" title="WizardPanel View 4">
+		</WizardPanels.Panel>
+		<WizardPanels.Panel footer="Footer in View 4" subtitle="A subtitle for View 4" title="WizardPanels View 4">
 			<div id="view4" />
 			<BodyText>A simple view</BodyText>
 			<buttons>
 				<Button>OK</Button>
 				<Button>Cancel</Button>
 			</buttons>
-		</View>
-	</WizardPanel>
+		</WizardPanels.Panel>
+	</WizardPanels>
 </div>;
 
 export default ThemeDecorator(app);