+++ CHANGELOG.md
@@ -2,14 +2,26 @@
 
 The following is a curated list of changes in the Enact sandstone module, newest changes on the top.
 
-## [unreleased]
+## [1.0.0-beta.2] - 2020-05-04
+
+### Deprecated
+
+- `sandstone/Panels.WizardPanel` and `sandstone/Panels.View`, replaced with `sandstone/WizardPanels` and `sandstone/WizardPanels.Panel` respectively
+- `sandstone/Panels.FlexiblePopupPanels`, replaced with `sandstone/FlexiblePopupPanels`
+- `sandstone/Panels.FixedPopupPanels`, replaced with `sandstone/FixedPopupPanels`
 
 ### Added
-- `sandstone/Scroller` and `sandstone/VirtualList` prop `initialHiddenHeight` to provide the initially hidden height of the vertical scrollbar when `featureContent` prop in the panel set to true.
+
+- `sandstone/WizardPanels` props `nextButtonAriaLabel`, `prevButtonAriaLabel`, `noNextButton`, `noPrevButton`, and `noSteps`
+- `sandstone/Scroller` and `sandstone/VirtualList` prop `initialHiddenHeight` to provide the height of the vertical scrollbar when the `featureContent` prop in the panel is set to true
+- `sandstone/Input.InputPopup` component
 
 ### Fixed
 
+- `sandstone/Header` centering
 - `sandstone/Input.InputField` disabled colors
+- `sandstone/WizardPanels` to hide previous and next buttons appropriately
+- `sandstone/TabLayout` to support disabled tabs
 
 ## [1.0.0-beta.1] - 2020-04-27